requirements.txt
numpy
pillow
tensorflow
scikit-learn
keras
opencv-python